refactor bibframe node definitions into separate file
