+++ lib/definitions/index.js
@@ -1,50 +1,20 @@
 import { compose, createFactory } from './utils'
 import activityStreamFactories from './activity-streams'
+import bibframeFactories from './bibframe'
 import StandardBehavior from './standard-behavior'
 import ContainerBehavior from './container-behavior'
 import ListItemBehavior from './list-item-behavior'
-import AgentBehavior from './foaf/agent-behavior'
-import PersonBehavior from './foaf/person-behavior'
-import WorkBehavior from './bibframe/work-behavior'
-import InstanceBehavior from './bibframe/instance-behavior'
-import ItemBehavior from './bibframe/item-behavior'
-import SharedBehavior from './bibframe/shared-behavior'
-import WorkOrInstanceBehavior from './bibframe/work-or-instance-behavior'
-import WorkInstanceOrItemBehavior from './bibframe/work-instance-or-item-behavior'
-import ContributionBehavior from './bibframe/contribution-behavior'
 
 const createStandardNode = createFactory(StandardBehavior)
 const createContainerNode = createFactory(compose(StandardBehavior, ContainerBehavior))
 const createListItemNode = createFactory(compose(StandardBehavior, ListItemBehavior))
 
-const WorkNode = createFactory(compose(
-  StandardBehavior,
-  SharedBehavior,
-  WorkInstanceOrItemBehavior,
-  WorkOrInstanceBehavior,
-  WorkBehavior
-))
-const InstanceNode = createFactory(compose(
-  StandardBehavior,
-  SharedBehavior,
-  WorkInstanceOrItemBehavior,
-  WorkOrInstanceBehavior,
-  InstanceBehavior
-))
-const ItemNode = createFactory(compose(
-  StandardBehavior,
-  SharedBehavior,
-  WorkInstanceOrItemBehavior,
-  ItemBehavior
-))
-
-const AgentNode = createFactory(compose(StandardBehavior, SharedBehavior, AgentBehavior))
-const PersonNode = createFactory(compose(StandardBehavior, SharedBehavior, AgentBehavior, PersonBehavior))
-const ContributionNode = createFactory(compose(StandardBehavior, SharedBehavior, ContributionBehavior))
-
 const FACTORY_BY_TYPE = Object.assign(
   {
     // basic types for hyper-reader document
+    // TODO: move away from storing documents in the db
+    // in future we should use store files, and db should be
+    // just bibliographic, annotations, and relationship data
     'hr:root': createContainerNode,
     'hr:head': createContainerNode,
     'hr:body': createContainerNode,
@@ -54,37 +24,9 @@ const FACTORY_BY_TYPE = Object.assign(
     'doco:Title': createContainerNode,
     'po:Block': createContainerNode,
     'po:Inline': createContainerNode,
-    'co:ListItem': createListItemNode,
-
-    'bf:Work': WorkNode,
-    // subclasses of Work
-    'bf:Text': WorkNode,
-    'bf:Cartography': WorkNode,
-    'bf:Audio': WorkNode,
-    'bf:NotatedMusic': WorkNode,
-    'bf:NotatedMovement': WorkNode,
-    'bf:Dataset': WorkNode,
-    'bf:StillImage': WorkNode,
-    'bf:MovingImage': WorkNode,
-    'bf:Object': WorkNode,
-    'bf:Multimedia': WorkNode,
-    'bf:MixedMaterial': WorkNode,
-
-    'bf:Instance': InstanceNode,
-    // subclasses of Instances
-    'bf:Print': InstanceNode,
-    'bf:Manuscript': InstanceNode,
-    'bf:Archival': InstanceNode,
-    'bf:Tactile': InstanceNode,
-    'bf:Electronic': InstanceNode,
-
-    'bf:Item': ItemNode,
-    'bf:Contribution': ContributionNode,
-    'bf:Agent': AgentNode,
-    'foaf:Agent': AgentNode,
-    'bf:Person': PersonNode,
-    'foaf:Person': PersonNode
+    'co:ListItem': createListItemNode
   },
+  bibframeFactories,
   activityStreamFactories
 )
